docs: prepare moving over examples
