+++ .prettierignore
@@ -1,2 +1,3 @@
 dist
+.next
 pnpm-lock.yaml
