+++ website/package.json
@@ -26,6 +26,7 @@
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
     "react-rx": "workspace:*",
+    "react-rx-old": "npm:react-rx@2.1.3",
     "rxjs": "^7.8.1",
     "styled-components": "^6.1.11",
     "typescript": "5.4.5"
