+++ website/src/examples/FizzBuzz/FizzBuzzExample.tsx
@@ -1,12 +1,35 @@
-import {
-  map,
-  React,
-  ReactDOM,
-  rxComponent,
-  take,
+import * as React from 'react'
+import * as ReactDOM from 'react-dom'
+import * as RxJS from 'rxjs'
+import * as operators from 'rxjs/operators'
+
+const {of, from, concat, merge} = RxJS
+const {
   timer,
-} from '../_utils/globalScope'
+  interval,
+  throwError,
+  combineLatest,
+  Observable,
+} = RxJS
+
+const {map, filter, reduce, scan, tap} = operators
+const {concatMap, mergeMap, switchMap, mapTo} =
+  operators
+const {startWith, catchError, take} = operators
 //@endimport
+
+import {observableCallback} from 'observable-callback'
+import {
+  context,
+  elementRef,
+  forwardRef,
+  handler,
+  rxComponent,
+  state,
+  useAsObservable,
+  useMemoObservable,
+  useObservable,
+} from 'react-rx-old'
 
 const FizzBuzzExample = rxComponent(
   timer(0, 500).pipe(
