+++ website/src/examples/event-handlers/EventHandlersExample.tsx
@@ -1,12 +1,35 @@
-import {
-  handler,
-  map,
-  React,
-  ReactDOM,
-  rxComponent,
-  startWith,
-} from '../_utils/globalScope'
+import * as React from 'react'
+import * as ReactDOM from 'react-dom'
+import * as RxJS from 'rxjs'
+import * as operators from 'rxjs/operators'
+
+const {of, from, concat, merge} = RxJS
+const {
+  timer,
+  interval,
+  throwError,
+  combineLatest,
+  Observable,
+} = RxJS
+
+const {map, filter, reduce, scan, tap} = operators
+const {concatMap, mergeMap, switchMap, mapTo} =
+  operators
+const {startWith, catchError, take} = operators
 //@endimport
+
+import {observableCallback} from 'observable-callback'
+import {
+  context,
+  elementRef,
+  forwardRef,
+  handler,
+  rxComponent,
+  state,
+  useAsObservable,
+  useMemoObservable,
+  useObservable,
+} from 'react-rx-old'
 
 const STYLE: React.CSSProperties = {
   display: 'flex',
