--- website/src/examples/forward-ref/index.tsx
@@ -1,12 +1,0 @@
-import {Example} from '../../_pages/Examples/Examples'
-
-const fs = require('fs')
-
-export const ForwardRefExample: Example = {
-  id: 'forward-ref',
-  title: 'Forward Ref',
-  source: fs.readFileSync(
-    `${__dirname}/ForwardRefExample.tsx`,
-    'utf-8',
-  ),
-}
