+++ website/src/examples/passthrough-props/PassThroughPropsExample.jsx
@@ -1,15 +1,36 @@
 import {formatDistance} from 'date-fns'
+import * as React from 'react'
+import * as ReactDOM from 'react-dom'
+import * as RxJS from 'rxjs'
+import * as operators from 'rxjs/operators'
 
-import {
-  combineLatest,
-  map,
-  React,
-  ReactDOM,
-  rxComponent,
-  take,
+const {of, from, concat, merge} = RxJS
+const {
   timer,
-} from '../_utils/globalScope'
+  interval,
+  throwError,
+  combineLatest,
+  Observable,
+} = RxJS
+
+const {map, filter, reduce, scan, tap} = operators
+const {concatMap, mergeMap, switchMap, mapTo} =
+  operators
+const {startWith, catchError, take} = operators
 //@endimport
+
+import {observableCallback} from 'observable-callback'
+import {
+  context,
+  elementRef,
+  forwardRef,
+  handler,
+  rxComponent,
+  state,
+  useAsObservable,
+  useMemoObservable,
+  useObservable,
+} from 'react-rx-old'
 
 const {share} = operators
