+++ website/src/examples/shared-state/SharedState.example.tsx
@@ -1,15 +1,37 @@
 import {observableCallback} from 'observable-callback'
 import {CSSProperties} from 'react'
+import * as React from 'react'
+import * as ReactDOM from 'react-dom'
+import * as RxJS from 'rxjs'
+import * as operators from 'rxjs/operators'
 
-import {
-  map,
-  of,
-  React,
-  ReactDOM,
-  rxComponent,
-  scan,
-} from '../_utils/globalScope'
+const {of, from, concat, merge} = RxJS
+const {
+  timer,
+  interval,
+  throwError,
+  combineLatest,
+  Observable,
+} = RxJS
+
+const {map, filter, reduce, scan, tap} = operators
+const {concatMap, mergeMap, switchMap, mapTo} =
+  operators
+const {startWith, catchError, take} = operators
 //@endimport
+
+import {observableCallback} from 'observable-callback'
+import {
+  context,
+  elementRef,
+  forwardRef,
+  handler,
+  rxComponent,
+  state,
+  useAsObservable,
+  useMemoObservable,
+  useObservable,
+} from 'react-rx-old'
 
 const CONTAINER: CSSProperties = {
   display: 'flex',
