+++ website/src/examples/todo-app/TodoApp.example.tsx
@@ -1,22 +1,38 @@
 import {ChangeEvent, FormEvent} from 'react'
-import styled from 'styled-components'
+import {styled} from 'styled-components'
 
-import {
+import * as React from 'react'
+import * as ReactDOM from 'react-dom'
+import * as RxJS from 'rxjs'
+import * as operators from 'rxjs/operators'
+
+const {of, from, concat, merge} = RxJS
+const {
+  timer,
+  interval,
+  throwError,
   combineLatest,
-  filter,
-  handler,
-  map,
-  mapTo,
-  merge,
-  operators,
-  React,
-  ReactDOM,
-  rxComponent,
-  scan,
-  startWith,
-  tap,
-} from '../_utils/globalScope'
+  Observable,
+} = RxJS
+
+const {map, filter, reduce, scan, tap} = operators
+const {concatMap, mergeMap, switchMap, mapTo} =
+  operators
+const {startWith, catchError, take} = operators
 //@endimport
+
+import {observableCallback} from 'observable-callback'
+import {
+  context,
+  elementRef,
+  forwardRef,
+  handler,
+  rxComponent,
+  state,
+  useAsObservable,
+  useMemoObservable,
+  useObservable,
+} from 'react-rx-old'
 
 const {withLatestFrom} = operators
