--- website/src/examples/use-element/index.tsx
@@ -1,12 +1,0 @@
-import {Example} from '../../_pages/Examples/Examples'
-
-const fs = require('fs')
-
-export const UseElementExample: Example = {
-  id: 'use-element',
-  title: 'Use element',
-  source: fs.readFileSync(
-    `${__dirname}/UseElementExample.tsx`,
-    'utf-8',
-  ),
-}
