+++ website/src/examples/use-observable/search/SearchExample.tsx
@@ -1,17 +1,37 @@
 import {Observable} from 'rxjs'
 
-import {
-  filter,
-  map,
-  of,
-  operators,
-  React,
-  ReactDOM,
-  switchMap,
+import * as React from 'react'
+import * as ReactDOM from 'react-dom'
+import * as RxJS from 'rxjs'
+import * as operators from 'rxjs/operators'
+
+const {of, from, concat, merge} = RxJS
+const {
   timer,
-  useMemoObservable,
-} from '../../_utils/globalScope'
+  interval,
+  throwError,
+  combineLatest,
+  Observable,
+} = RxJS
+
+const {map, filter, reduce, scan, tap} = operators
+const {concatMap, mergeMap, switchMap, mapTo} =
+  operators
+const {startWith, catchError, take} = operators
 //@endimport
+
+import {observableCallback} from 'observable-callback'
+import {
+  context,
+  elementRef,
+  forwardRef,
+  handler,
+  rxComponent,
+  state,
+  useAsObservable,
+  useMemoObservable,
+  useObservable,
+} from 'react-rx-old'
 
 const {distinctUntilChanged, debounceTime} =
   operators
